@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
     <!-- SonarCloud needs this -->
     <ProjectGuid>{d32c4ee9-e827-467a-b116-8ef0ba08a11f}</ProjectGuid>
@@ -10,8 +10,6 @@
   <ItemGroup>
     <PackageReference Include="AltinnCore.ServiceLibrary" Version="10.0.0" />
     <PackageReference Include="JWTCookieAuthentication" Version="0.9.0-alpha" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
     <PackageReference Include="Serilog.Extensions.Hosting" Version="2.0.0" />
     <PackageReference Include="Serilog.Extensions.Logging" Version="2.0.2" />
     <PackageReference Include="Serilog.Settings.Configuration" Version="3.0.1" />
